Download bpmn file only once
