+++ app/routes/bpmn-files/bpmn-file.js
@@ -1,6 +1,7 @@
 import Route from '@ember/routing/route';
 import { service } from '@ember/service';
-import { keepLatestTask } from 'ember-concurrency';
+import { keepLatestTask, task, waitForProperty } from 'ember-concurrency';
+import generateBpmnDownloadUrl from 'frontend-openproceshuis/utils/bpmn-download-url';
 
 export default class BpmnFilesBpmnFileRoute extends Route {
   @service store;
@@ -8,9 +9,18 @@ export default class BpmnFilesBpmnFileRoute extends Route {
   async model() {
     let { id: fileId } = this.paramsFor('bpmn-files.bpmn-file');
 
+    const loadMetadataTaskInstance = this.loadBpmnFileTask.perform(fileId);
+    const loadedMetadata = this.loadBpmnFileTask.lastSuccesful?.value;
+    const loadDiagramTaskInstance = this.loadDiagramTask.perform(
+      loadMetadataTaskInstance
+    );
+    const loadedDiagram = this.loadDiagramTask.lastSuccessful?.value;
+
     return {
-      loadMetadataTaskInstance: this.loadBpmnFileTask.perform(fileId),
-      loadedMetadata: this.loadBpmnFileTask.lastSuccesful?.value,
+      loadMetadataTaskInstance,
+      loadedMetadata,
+      loadDiagramTaskInstance,
+      loadedDiagram,
     };
   }
 
@@ -19,5 +29,16 @@ export default class BpmnFilesBpmnFileRoute extends Route {
     return yield this.store.findRecord('file', fileId, {
       include: 'publisher',
     });
+  }
+
+  @task
+  *loadDiagramTask(loadMetadataTaskInstance) {
+    yield waitForProperty(loadMetadataTaskInstance, 'isFinished');
+    const fileId = loadMetadataTaskInstance.value.id;
+
+    const url = generateBpmnDownloadUrl(fileId);
+    const response = yield fetch(url);
+    if (!response.ok) throw Error(response.status);
+    return yield response.text();
   }
 }
