+++ app/routes/bpmn-files/bpmn-file/index.js
@@ -1,6 +1,5 @@
 import Route from '@ember/routing/route';
-import generateBpmnDownloadUrl from 'frontend-openproceshuis/utils/bpmn-download-url';
-import { keepLatestTask, task, waitForProperty } from 'ember-concurrency';
+import { keepLatestTask, waitForProperty } from 'ember-concurrency';
 import { service } from '@ember/service';
 
 export default class BpmnFilesBpmnFileIndexRoute extends Route {
@@ -12,14 +11,12 @@ export default class BpmnFilesBpmnFileIndexRoute extends Route {
   };
 
   async model(params) {
-    let { loadMetadataTaskInstance, loadedMetadata } = this.modelFor(
-      'bpmn-files.bpmn-file'
-    );
-
-    let loadDiagramTaskInstance = this.loadDiagramTask.perform(
-      loadMetadataTaskInstance
-    );
-    let loadedDiagram = this.loadDiagramTask.lastSuccessful?.value;
+    let {
+      loadMetadataTaskInstance,
+      loadedMetadata,
+      loadDiagramTaskInstance,
+      loadedDiagram,
+    } = this.modelFor('bpmn-files.bpmn-file');
 
     let loadBpmnElementsTaskInstance = this.loadBpmnElementsTask.perform(
       loadMetadataTaskInstance,
@@ -35,17 +32,6 @@ export default class BpmnFilesBpmnFileIndexRoute extends Route {
       loadBpmnElementsTaskInstance,
       loadedBpmnElements,
     };
-  }
-
-  @task
-  *loadDiagramTask(loadMetadataTaskInstance) {
-    yield waitForProperty(loadMetadataTaskInstance, 'isFinished');
-    const fileId = loadMetadataTaskInstance.value.id;
-
-    const url = generateBpmnDownloadUrl(fileId);
-    const response = yield fetch(url);
-    if (!response.ok) throw Error(response.status);
-    return yield response.text();
   }
 
   @keepLatestTask({ cancelOn: 'deactivate' })
